moving getAllPaths and getContentPaths to scripts
