+++ docs/scripts/getContentPaths.ts
