+++ docs/src/data/pages.preval.ts
@@ -1,4 +1,4 @@
-import { getContentPaths } from '@/utils/getContentPaths';
+import { getContentPaths } from 'scripts/getContentPaths';
 import { getPagesManifest } from '@/utils/getPagesManifest';
 import { getPageFromSlug } from '@/utils/getPageFromSlug';
 import preval from 'next-plugin-preval';
